Removed unused trix code and move part of it into a safe navigating controller
